Removed euler namespace and moved files and functions to usecases namespace
