+++ include/physics_backend.hpp
@@ -1,7 +1,7 @@
 #ifndef PHYSICS_BACKEND_HPP
 #define PHYSICS_BACKEND_HPP
 
-#include "physics_backend/euler.hpp"
+#include "physics_backend/simulation.hpp"
 #include "physics_backend/units.hpp"
 
 #endif // PHYSICS_BACKEND_HPP
