--- include/physics_backend/euler.hpp
@@ -1,7 +1,0 @@
-#ifndef PHYSICS_BACKEND_EULER_HPP
-#define PHYSICS_BACKEND_EULER_HPP
-
-#include "physics_backend/usecases/euler/particle.hpp"
-#include "physics_backend/usecases/euler/step.hpp"
-
-#endif // PHYSICS_BACKEND_EULER_HPP
