+++ include/physics_backend/simulation.hpp
@@ -1,0 +1,7 @@
+#ifndef PHYSICS_BACKEND_SIMULATION_HPP
+#define PHYSICS_BACKEND_SIMULATION_HPP
+
+#include "physics_backend/usecases/particle.hpp"
+#include "physics_backend/usecases/step.hpp"
+
+#endif // PHYSICS_BACKEND_SIMULATION_HPP
