+++ include/physics_backend/usecases/step.hpp
@@ -1,5 +1,5 @@
-#ifndef PHYSICS_BACKEND_USECASES_EULER_STEP_HPP
-#define PHYSICS_BACKEND_USECASES_EULER_STEP_HPP
+#ifndef PHYSICS_BACKEND_USECASES_STEP_HPP
+#define PHYSICS_BACKEND_USECASES_STEP_HPP
 
 #include <ranges>
 
@@ -8,10 +8,10 @@
 #include "physics_backend/usecases/constrain.hpp"
 #include "physics_backend/usecases/polygon.hpp"
 #include "physics_backend/usecases/detail/fmaps.hpp"
-#include "physics_backend/usecases/euler/motion.hpp"
-#include "physics_backend/usecases/euler/particle.hpp"
+#include "physics_backend/usecases/motion.hpp"
+#include "physics_backend/usecases/particle.hpp"
 
-namespace physics::euler
+namespace physics::usecases
 {
 
 template<
@@ -20,7 +20,7 @@ template<
     physics::units::IsTimeUnit Time>
 auto step(
     Container<Particle<Units>> const& particles,
-    physics::usecases::Polygon2D<typename Units::Length> const& constraint,
+    Polygon2D<typename Units::Length> const& constraint,
     Time time,
     uint32_t substeps = 1)
 {
@@ -36,17 +36,17 @@ auto step(
     for(size_t substep {0}; substep < substeps; ++substep)
     {
         auto motion = [&time](auto const& particle){ return resolveMotion(particle, time); };
-        auto constrain = [&constraint](auto const& particle){ return physics::usecases::resolveConstraint(particle, constraint); };
+        auto constrain = [&constraint](auto const& particle){ return resolveConstraint(particle, constraint); };
 
         auto updatedParticles {physics::detail::fmaps(wrappedParticles, motion, constrain)};
         wrappedParticles = Container<Particle<Units>>(updatedParticles.begin(), updatedParticles.end());
 
-        physics::usecases::resolveCollisions(wrappedParticles);
+        resolveCollisions(wrappedParticles);
     }
 
     return wrappedParticles;
 }
 
-} // namespace physics::euler
+} // namespace physics::usecases
 
-#endif // PHYSICS_BACKEND_USECASES_EULER_STEP_HPP
+#endif // PHYSICS_BACKEND_USECASES_STEP_HPP
