+++ test/usecases/CMakeLists.txt
@@ -1,9 +1,9 @@
 target_sources(test_physics_backend
     PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/test_polygon.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/test_constrain.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/test_collision.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/test_polygon.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/test_simulation.cpp
 )
 
-add_subdirectory(euler)
 add_subdirectory(detail)
