--- test/usecases/euler/CMakeLists.txt
@@ -1,4 +1,0 @@
-target_sources(test_physics_backend
-    PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/test_euler.cpp
-)
