+++ test/usecases/test_constrain.cpp
@@ -2,7 +2,7 @@
 
 #include "common/precision.hpp"
 #include "physics_backend/domain/vector.hpp"
-#include "physics_backend/euler.hpp"
+#include "physics_backend/simulation.hpp"
 #include "physics_backend/usecases/constrain.hpp"
 
 TEST(ConstrainTest, CenterInPeekingOut)
@@ -19,7 +19,7 @@ TEST(ConstrainTest, CenterInPeekingOut)
         }
     };
 
-    physics::euler::Particle<physics::units::SI> particle {
+    physics::usecases::Particle<physics::units::SI> particle {
         .mass {1.0_kg},
         .radius {0.25_m},
         .position {physics::domain::PositionVector2D(0.1_m, 0.05_m)},
@@ -48,7 +48,7 @@ TEST(ConstrainTest, SingleEdgeReflection)
         }
     };
 
-    physics::euler::Particle<physics::units::SI> particle {
+    physics::usecases::Particle<physics::units::SI> particle {
         .mass {1.0_kg},
         .radius {0.25_m},
         .position {physics::domain::PositionVector2D(0.1_m, 5.0_m)},
@@ -76,7 +76,7 @@ TEST(ConstrainTest, InBounds)
         }
     };
 
-    physics::euler::Particle<physics::units::SI> particle {
+    physics::usecases::Particle<physics::units::SI> particle {
         .mass {1.0_kg},
         .radius {0.25_m},
         .position {physics::domain::PositionVector2D(0.5_m, 0.5_m)},
@@ -91,34 +91,4 @@ TEST(ConstrainTest, InBounds)
 
     auto expectedVelocity {physics::domain::Vector2D::fromComponents(1.0_mps, -1.0_mps)};
     EXPECT_TRUE(physics::domain::Vector2D::compare(updatedParticle.velocity, expectedVelocity, physics::test::REQUIRED_PRECISION));
-}
-
-TEST(ConstrainTest, Tunneling)
-{
-    // TODO: Solve the tunneling problem.
-
-    // using namespace physics::units::literals;
-    // using Length = physics::units::SI::Length;
-
-    // physics::usecases::Polygon2D<Length> constraint {
-    //     {
-    //         physics::domain::PositionVector2D(0.0_m, 0.0_m),
-    //         physics::domain::PositionVector2D(1.0_m, 0.0_m),
-    //         physics::domain::PositionVector2D(1.0_m, 1.0_m),
-    //         physics::domain::PositionVector2D(0.0_m, 1.0_m)
-    //     }
-    // };
-
-    // physics::euler::Particle<physics::units::SI> particle {
-    //     .mass {1.0_kg},
-    //     .radius {0.25_m},
-    //     .position {physics::domain::PositionVector2D(-1.0_m, -1.0_m)},
-    //     .velocity {physics::domain::VelocityVector2D(0.0_mps, 0.0_mps)},
-    //     .forces {physics::domain::ForceVector2D(0.0_N, -9.81_N)}
-    // };
-
-    // auto updatedParticle {physics::usecases::resolveConstraint(particle, constraint)};
-    // auto expectedPosition {physics::domain::PositionVector2D(0.5_m, 0.5_m)};
-
-    // EXPECT_TRUE(physics::domain::Vector2D::compare(updatedParticle.position, expectedPosition, physics::test::REQUIRED_PRECISION));
 }
